added points to each alien
