+++ Main.js
@@ -1,9 +1,15 @@
+import { Player } from './classes/Player.js';
+import { Alien } from './classes/Alien.js';
+import { Bullet } from './classes/Bullet.js';
+import { Heart } from './classes/Heart.js';
+import { Particle } from './classes/Particle.js';
+
 const canvas = document.getElementById("gameCanvas");
 const ctx = canvas.getContext("2d");
 const scoreElement = document.getElementById("score");
 const levelElement = document.getElementById("level");
 const livesElement = document.getElementById("lives");
-startButton = document.getElementById("startButton");
+const startButton = document.getElementById("startButton");
 const restartButton = document.getElementById("restartButton");
 const gameOverElement = document.getElementById("gameOver");
 const highScoreElement = document.getElementById("highScoreValue");
@@ -63,6 +69,14 @@ function updateScore() {
       document.getElementById('highScore').textContent = highScore; // Update the high score display
   }
 }
+function gameLoop() {
+  if (!gamePaused) {
+      update(); // Update game objects like player, aliens, hearts
+      draw();    // Draw everything to the canvas
+  }
+  requestAnimationFrame(gameLoop); // Request the next frame
+}
+
 function resumeGame() {
     gamePaused = false; // Set the game paused state to false
     console.log("Game resumed"); // Log resume action (optional)
@@ -92,6 +106,7 @@ backgroundMusic.volume = volumeSlider.value;
 
 startButton.addEventListener("click", function () {
   backgroundMusic.volume = volumeSlider.value;  // Set volume when game starts
+  startGame();
 });
 
 // Drop down menu event listeners
@@ -134,103 +149,6 @@ function setLevel(difficulty){
   }
 }
 
-class Player {
-  constructor() {
-    this.width = 60;
-    this.height = 60;
-    this.x = canvas.width / 2 - this.width / 2;
-    this.y = canvas.height - this.height - 10;
-    this.speed = 5;
-  }
-
-  draw() {
-    ctx.fillStyle = "#4a4a4a";
-    ctx.beginPath();
-    ctx.moveTo(this.x + this.width / 2, this.y);
-    ctx.lineTo(this.x, this.y + this.height);
-    ctx.lineTo(this.x + this.width, this.y + this.height);
-    ctx.closePath();
-    ctx.fill();
-
-    ctx.fillStyle = "#00ffff";
-    ctx.beginPath();
-    ctx.arc(
-      this.x + this.width / 2,
-      this.y + this.height / 3,
-      this.width / 6,
-      0,
-      Math.PI * 2
-    );
-    ctx.fill();
-
-    ctx.fillStyle = "#ff0000";
-    ctx.fillRect(this.x - 10, this.y + this.height - 20, 10, 20);
-    ctx.fillRect(this.x + this.width, this.y + this.height - 20, 10, 20);
-  }
-
-  move() {
-    if ((keys.ArrowLeft || keys["KeyA"]) && this.x > 0)
-      this.x -= this.speed;
-    if ((keys.ArrowRight || keys["KeyD"]) && this.x < canvas.width - this.width)
-      this.x += this.speed;
-  }
-}
-
-class Heart {
-  constructor(x, y) {
-    this.width = 30;
-    this.height = 30;
-    this.x = x;
-    this.y = y;
-    this.speed = 2; 
-  }
-
-  draw() {
-    ctx.fillStyle = "#FF0000";
-    
-    // Heart shape using bezier curves
-    ctx.beginPath();
-    const topCurveHeight = this.height * 0.3;
-    ctx.moveTo(this.x, this.y + topCurveHeight);
-    
-    // Top left curve
-    ctx.bezierCurveTo(
-      this.x, this.y, 
-      this.x - this.width / 2, this.y, 
-      this.x - this.width / 2, this.y + topCurveHeight
-    );
-    
-    // Bottom left curve
-    ctx.bezierCurveTo(
-      this.x - this.width / 2, this.y + (this.height + topCurveHeight) / 2, 
-      this.x, this.y + (this.height + topCurveHeight) / 2, 
-      this.x, this.y + this.height
-    );
-    
-    // Bottom right curve
-    ctx.bezierCurveTo(
-      this.x, this.y + (this.height + topCurveHeight) / 2, 
-      this.x + this.width / 2, this.y + (this.height + topCurveHeight) / 2, 
-      this.x + this.width / 2, this.y + topCurveHeight
-    );
-
-    // Top right curve
-    ctx.bezierCurveTo(
-      this.x + this.width / 2, this.y, 
-      this.x, this.y, 
-      this.x, this.y + topCurveHeight
-    );
-    
-    ctx.closePath();
-    ctx.fill();
-  }
-
-  move() {
-    this.y += this.speed;
-  }
-}
-
-
 let hearts = [];
 
 function spawnHearts() {
@@ -239,299 +157,8 @@ function spawnHearts() {
   }
 }
 
-class Alien {
-  constructor(x, y, type) {
-    this.width = 40;
-    this.height = 40;
-    this.x = x;
-    this.y = y;
-    this.speed = 1 + level * 0.5;
-    this.type = type; // Assign the type
-  }
 
-  draw() {
-    if (this.type === 'default') {
-      ctx.fillStyle = "#32a852"; // Alien green color for the body
-      ctx.beginPath();
-      ctx.arc(
-        this.x + this.width / 2, 
-        this.y + this.height / 2, 
-        this.width / 2, 
-        0,
-        Math.PI * 2 // Full circle
-      );
-      ctx.fill();
-  
-      // Eyes
-      ctx.fillStyle = "#ffffff"; 
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.beginPath();
-      ctx.arc(this.x + (2 * this.width) / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-      ctx.fill();
-  
-      // Pupils
-      ctx.fillStyle = "#000000";
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 12, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.beginPath();
-      ctx.arc(this.x + (2 * this.width) / 3, this.y + this.height / 3, this.width / 12, 0, Math.PI * 2);
-      ctx.fill();
-  
-      // Antennae
-      ctx.strokeStyle = "#ff00ff"; 
-      ctx.lineWidth = 2;
-      ctx.beginPath();
-      ctx.moveTo(this.x + this.width / 3, this.y); 
-      ctx.lineTo(this.x + this.width / 3, this.y - this.height / 4);
-      ctx.stroke();
-  
-      ctx.beginPath();
-      ctx.moveTo(this.x + (2 * this.width) / 3, this.y);
-      ctx.lineTo(this.x + (2 * this.width) / 3, this.y - this.height / 4);
-      ctx.stroke();
-    } 
-    else if (this.type === 'terrific') {
-      // Terrific Alien appearance
-      ctx.fillStyle = "#8b0000"; // Dark red body
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 2, this.y + this.height / 2, this.width / 2, 0, Math.PI * 2);
-      ctx.fill();
 
-      // Eyes (one misaligned pupil to look scary)
-      ctx.fillStyle = "#ffffff";
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-      ctx.fill();
-
-      ctx.beginPath();
-      ctx.arc(this.x + (2 * this.width) / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-      ctx.fill();
-
-      // Pupils (misaligned one)
-      ctx.fillStyle = "#000000";
-      // Left pupil
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 12, 0, Math.PI * 2);
-      ctx.fill();
-
-      // Right pupil (slightly offset)
-      ctx.beginPath();
-      ctx.arc(this.x + (2 * this.width) / 3 + 5, this.y + this.height / 3 + 5, this.width / 12, 0, Math.PI * 2);
-      ctx.fill();
-
-      // Add horns
-      ctx.strokeStyle = "#8b0000"; 
-      ctx.lineWidth = 3;
-      ctx.beginPath();
-      ctx.moveTo(this.x + this.width / 4, this.y); 
-      ctx.lineTo(this.x + this.width / 5, this.y - this.height / 4); // Left horn
-      ctx.stroke();
-
-      ctx.beginPath();
-      ctx.moveTo(this.x + (3 * this.width) / 4, this.y); 
-      ctx.lineTo(this.x + (4 * this.width) / 5, this.y - this.height / 4); // Left horn
-      ctx.stroke();
-
-      // Add jagged teeth
-      ctx.fillStyle = "#ffffff";
-      ctx.beginPath();
-      ctx.moveTo(this.x + this.width / 4, this.y + (2 * this.height) / 3); // Left tooth
-      ctx.lineTo(this.x + this.width / 2, this.y + (3 * this.height) / 4); // Middle tooth
-      ctx.lineTo(this.x + (3 * this.width) / 4, this.y + (2 * this.height) / 3); // Right tooth
-      ctx.closePath();
-      ctx.fill();
-
-      // Add scar (diagonal line across face)
-      ctx.strokeStyle = "#ff0000";
-      ctx.lineWidth = 3;
-      ctx.beginPath();
-      ctx.moveTo(this.x + this.width / 4, this.y + this.height / 2);
-      ctx.lineTo(this.x + (3 * this.width) / 4, this.y + this.height / 3);
-      ctx.stroke();
-    }
-    else if (this.type === 'cute') {
-      // Cute Alien appearance
-      ctx.fillStyle = "#FF69B4"; 
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 2, this.y + this.height / 2, this.width / 2, 0, Math.PI * 2);
-      ctx.fill();
-      
-      // Big round eyes
-      ctx.fillStyle = "#ffffff"; 
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 4, this.y + this.height / 2.5, this.width / 4, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.beginPath();
-      ctx.arc(this.x + (3 * this.width) / 4, this.y + this.height / 2.5, this.width / 4, 0, Math.PI * 2);
-      ctx.fill();
-      
-      // Big black pupils
-      ctx.fillStyle = "#000000";
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 4, this.y + this.height / 2.5, this.width / 8, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.beginPath();
-      ctx.arc(this.x + (3 * this.width) / 4, this.y + this.height / 2.5, this.width / 8, 0, Math.PI * 2);
-      ctx.fill();
-      
-      // Cute blushing cheeks (small pink circles)
-      ctx.fillStyle = "#FFC0CB";
-      ctx.beginPath();
-      ctx.arc(this.x + this.width / 4, this.y + (2 * this.height) / 3, this.width / 8, 0, Math.PI * 2);
-      ctx.fill();
-      ctx.beginPath();
-      ctx.arc(this.x + (3 * this.width) / 4, this.y + (2 * this.height) / 3, this.width / 8, 0, Math.PI * 2);
-      ctx.fill();
-
-      ctx.fillStyle = "#FF0000";
-      ctx.fillStyle = "#FF0000";
-      // Horns
-      ctx.strokeStyle = "#FF69B4";
-      ctx.lineWidth = 2;
-      ctx.beginPath();
-      ctx.moveTo(this.x + this.width / 4, this.y);
-      ctx.lineTo(this.x + this.width / 5, this.y - this.height / 4); // Left horn
-      ctx.stroke();
-
-      ctx.beginPath();
-      ctx.moveTo(this.x + (3 * this.width) / 4, this.y);
-      ctx.lineTo(this.x + (4 * this.width) / 5, this.y - this.height / 4); // Right horn
-      ctx.stroke();
-      
-    } 
-    else if (this.type === 'robotic') {
-      // Robotic Alien appearance
-      ctx.fillStyle = "#808080"; // Gray for metal body
-        ctx.beginPath();
-        ctx.arc(this.x + this.width / 2, this.y + this.height / 2, this.width / 2, 0, Math.PI * 2);
-        ctx.fill();
-
-        // Eyes
-        ctx.fillStyle = "#ff0000"; // Red robotic eyes
-        ctx.beginPath();
-        ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-        ctx.fill();
-
-        ctx.beginPath();
-        ctx.arc(this.x + (2 * this.width) / 3, this.y + this.height / 3, this.width / 6, 0, Math.PI * 2);
-        ctx.fill();
-
-        // Pupils
-        ctx.fillStyle = "#000000";
-        ctx.beginPath();
-        ctx.arc(this.x + this.width / 3, this.y + this.height / 3, this.width / 12, 0, Math.PI * 2);
-        ctx.fill();
-
-        ctx.beginPath();
-        ctx.arc(this.x + (2 * this.width) / 3, this.y + this.height / 3, this.width / 12, 0, Math.PI * 2);
-        ctx.fill();
-
-        // antennae with lights
-        ctx.strokeStyle = "#ff0000"; 
-        ctx.lineWidth = 3;
-        // Left antenna
-        ctx.beginPath();
-        ctx.moveTo(this.x + this.width / 3, this.y);
-        ctx.lineTo(this.x + this.width / 3, this.y - this.height / 4);
-        ctx.stroke();
-
-        // Red light on top
-        ctx.fillStyle = "#ff0000";
-        ctx.beginPath();
-        ctx.arc(this.x + this.width / 3, this.y - this.height / 4 - 5, 5, 0, Math.PI * 2);
-        ctx.fill();
-
-        // Right antenna
-        ctx.beginPath();
-        ctx.moveTo(this.x + (2 * this.width) / 3, this.y);
-        ctx.lineTo(this.x + (2 * this.width) / 3, this.y - this.height / 4);
-        ctx.stroke();
-
-        // Red light on top
-        ctx.fillStyle = "#ff0000";
-        ctx.beginPath();
-        ctx.arc(this.x + (2 * this.width) / 3, this.y - this.height / 4 - 5, 5, 0, Math.PI * 2);
-        ctx.fill();
-
-        // mouth
-        ctx.fillStyle = "#000000";
-        ctx.fillRect(this.x + this.width / 3, this.y + (2 * this.height) / 3, this.width / 3, this.height / 6);
-    } 
-    else if (this.type === 'ghostly') {
-      //body
-        ctx.fillStyle = "#F8F8FF"; 
-        ctx.beginPath();
-        ctx.arc(this.x + this.width / 2, this.y + this.height / 2, this.width / 2, 0, Math.PI * 2);
-        ctx.fill();
-        
-        // Eyes 
-        ctx.fillStyle = "#000000";
-        ctx.beginPath();
-        ctx.ellipse(this.x + this.width / 4, this.y + this.height / 3, this.width / 8, this.height / 4, 0, 0, Math.PI * 2);
-        ctx.fill();
-        ctx.beginPath();
-        ctx.ellipse(this.x + (3 * this.width) / 4, this.y + this.height / 3, this.width / 8, this.height / 4, 0, 0, Math.PI * 2);
-        ctx.fill();
-        
-        // Tail
-        ctx.fillStyle = "#F8F8FF";
-        ctx.beginPath();
-        ctx.moveTo(this.x, this.y + this.height);
-        ctx.quadraticCurveTo(this.x + this.width / 2, this.y + this.height + 10, this.x + this.width, this.y + this.height);
-        ctx.fill();
-    }
-  }
-
-  move() {
-    this.y += this.speed;
-  }
-}
-
-class Bullet {
-  constructor(x, y) {
-    this.width = 5;
-    this.height = 15;
-    this.x = x;
-    this.y = y;
-    this.speed = 7;
-  }
-
-  draw() {
-    ctx.fillStyle = "#0ff";
-    ctx.fillRect(this.x, this.y, this.width, this.height);
-  }
-
-  move() {
-    this.y -= this.speed;
-  }
-}
-
-class Particle {
-  constructor(x, y) {
-    this.x = x;
-    this.y = y;
-    this.size = Math.random() * 5 + 1;
-    this.speedX = Math.random() * 4 - 2;
-    this.speedY = Math.random() * 4 - 2;
-    this.color = `hsl(${Math.random() * 360}, 100%, 50%)`;
-  }
-
-  draw() {
-    ctx.fillStyle = this.color;
-    ctx.beginPath();
-    ctx.arc(this.x, this.y, this.size, 0, Math.PI * 2);
-    ctx.fill();
-  }
-
-  update() {
-    this.x += this.speedX;
-    this.y += this.speedY;
-    if (this.size > 0.2) this.size -= 0.1;
-  }
-}
 
 function initGame() {
   player = new Player();
@@ -548,26 +175,32 @@ function initGame() {
   spawnAliens();
 }
 
+// spawn only 10 aliens
 function spawnAliens() {
-  // Spawn only 10 aliens for each wave
   if (aliens.length >= totalAliens) return; 
 
   for (let i = 0; i < totalAliens; i++) {
     const alienType = Math.random();
     let type;
+    let points;
+
     if (alienType < 1 / 5) {
       type = "default";
+      points = 10; // Points for default aliens
     } else if (alienType < 2 / 5) {
       type = "terrific";
+      points = 20; // Points for terrific aliens
     } else if (alienType < 3 / 5) {
       type = "cute";
+      points = 15; // Points for cute aliens
     } else if (alienType < 4 / 5) {
       type = "robotic";
+      points = 25; // Points for robotic aliens
     } else {
       type = "ghostly";
+      points = 30; // Points for ghostly aliens
     }
-
-    // Spawn aliens randomly
+   // spawn aliens randomly
     const numAliens = Math.floor(Math.random() * 3) + 2;
 
     for (let j = 0; j < numAliens && aliens.length < totalAliens; j++) {
@@ -575,12 +208,14 @@ function spawnAliens() {
         new Alien(
           Math.random() * (canvas.width - 40), 
           -50 - Math.random() * 500, 
-          type
+          type,
+          points // Pass points to the alien constructor
         )
       );
     }
   }
 }
+
 
 function update() {
   if (!gameActive || gamePaused || waitingForNextWave) return; 
@@ -653,30 +288,29 @@ function update() {
     }
 
     bullets.forEach((bullet, bulletIndex) => {
-      if (
-        bullet.x < alien.x + alien.width &&
-        bullet.x + bullet.width > alien.x &&
-        bullet.y < alien.y + alien.height &&
-        bullet.y + bullet.height > alien.y
-      ) {
-        // Destroy the alien
-        aliens.splice(alienIndex, 1);
-        bullets.splice(bulletIndex, 1);
-      
-        // Increment the destroyed aliens count
-        aliensDestroyed++;
-        score++;
-        scoreElement.textContent = score;
-        hitSound.currentTime = 0;
-        hitSound.play();
-      
-        // Check if all 10 aliens have been destroyed
-        if (aliensDestroyed >= totalAliens) {
-          endWave();
-        }
+  if (
+    bullet.x < alien.x + alien.width &&
+    bullet.x + bullet.width > alien.x &&
+    bullet.y < alien.y + alien.height &&
+    bullet.y + bullet.height > alien.y
+  ) {
+    // Destroy the alien and get points
+    aliens.splice(alienIndex, 1);
+    bullets.splice(bulletIndex, 1);
+    
+    // Increment the destroyed aliens count and score based on alien type
+    score += alien.points; // Add points from the defeated alien
+    scoreElement.textContent = score; // Update score display
+    hitSound.currentTime = 0;
+    hitSound.play();
+    
+    // Check if all aliens have been destroyed
+    if (aliensDestroyed >= totalAliens) {
+      endWave();
       }
-    });
+    }
   });
+});
 
   bullets.forEach((bullet, bulletIndex) => {
     bullet.draw();
