+++ index.html
@@ -179,7 +179,7 @@
     </div>
 
     <!-- Scripts -->
-    <script src="Alien.js"></script>
+    <script src="Main.js"></script>
     <script src="darkMode.js"></script>
 
     <script>
